requirements.txt
anthropic>=0.40.0
supabase>=2.4.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
python-docx>=1.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
apscheduler>=3.10.4
